+++ smartsim/_core/control/controller.py
@@ -106,6 +106,7 @@ class Controller:
         manifest: Manifest,
         block: bool = True,
         kill_on_interrupt: bool = True,
+        monitor: bool = True,
     ) -> None:
         """Start the passed SmartSim entities
 
@@ -121,7 +122,7 @@ class Controller:
         SignalInterceptionStack.get(signal.SIGINT).push_unique(
             self._jobs.signal_interrupt
         )
-        self._launch(exp_name, exp_path, manifest)
+        self._launch(exp_name, exp_path, manifest, monitor)
 
         # start the job manager thread if not already started
         if not self._jobs.actively_monitoring:
@@ -155,7 +156,7 @@ class Controller:
         :param kill_on_interrupt: flag for killing jobs when SIGINT is received
         """
         self._jobs.kill_on_interrupt = kill_on_interrupt
-        to_monitor = self._jobs.jobs
+        to_monitor = self._jobs.monitor_jobs
         while len(to_monitor) > 0:
             time.sleep(interval)
 
@@ -370,7 +371,9 @@ class Controller:
                 "Symlinking files failed."
             )
 
-    def _launch(self, _exp_name: str, exp_path: str, manifest: Manifest) -> None:
+    def _launch(
+        self, _exp_name: str, exp_path: str, manifest: Manifest, monitor: bool = True
+    ) -> None:
         """Main launching function of the controller
 
         Orchestrators are always launched first so that the
@@ -379,6 +382,7 @@ class Controller:
         :param exp_name: The name of the launching experiment
         :param exp_path: path to location of ``Experiment`` directory if generated
         :param manifest: Manifest of deployables to launch
+        :param monitor: boolean to signal whether to monitor deployables
         """
 
         # Create a unique timestamp for this launch to ensure unique metadata
@@ -454,7 +458,7 @@ class Controller:
 
         # launch and symlink steps
         for step, entity in steps:
-            self._launch_step(step, entity)
+            self._launch_step(step, entity, monitor)
             self.symlink_output_files(step, entity)
 
         # symlink substeps to maintain directory structure
@@ -533,11 +537,13 @@ class Controller:
         self,
         job_step: Step,
         entity: SmartSimEntity | EntitySequence[SmartSimEntity],
+        monitor: bool = True,
     ) -> None:
         """Use the launcher to launch a job step
 
         :param job_step: a job step instance
         :param entity: entity instance
+        :param monitor: boolean determining whether to monitor job
         :raises SmartSimError: if launch fails
         """
         # attempt to retrieve entity name in JobManager.completed
@@ -582,10 +588,10 @@ class Controller:
 
         if self._jobs.query_restart(entity.name):
             logger.debug(f"Restarting {entity.name}")
-            self._jobs.restart_job(job_step.name, job_id, entity.name, is_task)
+            self._jobs.restart_job(job_step.name, job_id, entity.name, is_task, monitor)
         else:
             logger.debug(f"Launching {entity.name}")
-            self._jobs.add_job(job_step.name, job_id, entity, is_task)
+            self._jobs.add_job(job_step.name, job_id, entity, is_task, monitor)
 
     def _create_batch_job_step(
         self,