+++ smartsim/_core/control/jobmanager.py
@@ -65,6 +65,7 @@ class JobManager:
 
         # active jobs
         self.jobs: dict[str, Job] = {}
+        self.monitor_jobs: dict[str, Job] = {}
         self.db_jobs: dict[str, Job] = {}
 
         # completed jobs
@@ -132,6 +133,8 @@ class JobManager:
                 del self.db_jobs[job.ename]
             elif job.ename in self.jobs:
                 del self.jobs[job.ename]
+                if job.ename in self.monitor_jobs:
+                    del self.monitor_jobs[job.ename]
 
     def __getitem__(self, entity_name: str) -> Job:
         """Return the job associated with the name of the entity
@@ -165,12 +168,14 @@ class JobManager:
         job_id: str | None,
         entity: SmartSimEntity | EntitySequence[SmartSimEntity],
         is_task: bool = True,
+        monitor: bool = True,
     ) -> None:
         """Add a job to the job manager which holds specific jobs by type.
 
         :param job_name: name of the job step
         :param job_id: job step id created by launcher
         :param entity: entity that was launched on job step
+        :param monitor: boolean to monitor job
         :param is_task: process monitored by TaskManager (True) or the WLM (True)
         """
         launcher = str(self._launcher)
@@ -180,6 +185,8 @@ class JobManager:
             self.db_jobs[entity.name] = job
         else:
             self.jobs[entity.name] = job
+            if monitor:
+                self.monitor_jobs[entity.name] = job
 
     def is_finished(self, entity: SmartSimEntity) -> bool:
         """Detect if a job has completed
@@ -264,6 +271,7 @@ class JobManager:
         job_id: str | None,
         entity_name: str,
         is_task: bool = True,
+        monitor: bool = True,
     ) -> None:
         """Function to reset a job to record history and be
         ready to launch again.
@@ -272,6 +280,7 @@ class JobManager:
         :param job_id: new job id
         :param entity_name: name of the entity of the job
         :param is_task: process monitored by TaskManager (True) or the WLM (True)
+        :param monitor: boolean to monitor job
 
         """
         with self._lock:
@@ -283,6 +292,8 @@ class JobManager:
                 self.db_jobs[entity_name] = job
             else:
                 self.jobs[entity_name] = job
+                if monitor:
+                    self.monitor_jobs[entity_name] = job
 
     def get_db_host_addresses(self) -> dict[str, list[str]]:
         """Retrieve the list of hosts for the database